+++ lib/native_glass_navbar.dart
@@ -2,6 +2,7 @@
 library native_glass_navbar;
 
 export 'liquid_glass_helper.dart';
+export 'native_glass_button.dart';
 
 import 'dart:async';
 import 'dart:developer' as developer;